Remove Logger 'message' Event Listeners
The `ConsoleLog` method of adds a new 'message' event listener to the
provided `Logger` instance. Repeated calls to `ConsoleLog` results in
Node warning of a possible memory leak. It also results in an
increasing number of duplicate log statements.

This change removes all 'message' event listeners of the given `Logger`
instance before adding the new one.

To reproduce: Start a new Zetta server. On a regular interval, stop all
`_peerClients` and call `listen` again.

(Add `this.setMaxListeners(2);` to `init` to shorten time to reproduce)