Handle Server Listen Error Event
If a port is already occupied, the `net` module will generate an
error event. If a callback is passed to the `ZettaHttpServer`
`listen` method, it will not be called, and the error will
go unhandled.

This update allows callers to `listen` to be able to properly handle
errors generated when starting the server.