Issue #351 - Removes Unecessary Event Handler
The `_subscribe` method adds a new "close" event handler to the socket
on every subscription. This results in Node warning of a possible
memory leak after 11 (default max listeners) subscriptions.

The event handler is not needed because one is added to the socket in
the `_streamEnabledClient` method. That method iterates through and
calls all subscription `unsubscribe` functions.

This commit removes the uncessary "close" event handler added on
every subscription.