Stream Enabled Clients Do Not Need Close Handler (#353)
The call to `EventBroker.prototype.client` checks the given `client`
instance for `streamEnabled`. If "true", the client is passed to
`_streamEnabledClient` where it has "close" and "unsubscribe" event
handlers added. If the client is *not* `streamEnabled` the event
handlers are *not* added.

This means that the `_subscribe` method only needs to make sure to
call unsubscribe on the "close" event of clients that are not
`streamEnabled`.