Support multiple versions of spdy transports within the z2z protocol.
Before the spdy-upgrade we defaulted to `spdy/3.1` when creating the
z2z protocol when the cloud -> hub after the websocket upgrade. This
allows the client to send which spdy/http2 protocol the cloud should
use to talk to the peer. In HTTP this is typically done using ALPN.

This allows new zetta versions to use `h2` while old clients will
default to `spdy/3.1`.