Zetta Version (1.6.0) - Upgraded spdy - Working with Node v18 (#379)
* First at upgrading spdy@3.x and ws@3.x

* Spdy/h2 ping working using forked version of `spdy`.

* Add extra upgrade header data back to socket for spdy to parse

* Fix virtual device  subscription causing socket hang up

* Update Travis to remove 0.10, add 6,8,10

* Changed custom spdy install to github:

* Updates to fix streams closing the connection.

* Fix isssue with reconnects not passing proper connectionId to peer client.
Fix issue with stream deduping when finding intenral remote port in the spdy internals.

* Fix failing test where an embedded ws is expected to return 404.

Updates to the WS cause it to return a 400 instead of a 404 when
the base path does not match.

* PeerSocket fix issue where a closing the peer conneciton causes subscribe's cb
to fire a second time. Wrap cb in a once pattern.

PeerSocket guard against this.ws not being initialized when closing the connection

* Add TODO to skipped test that tests spdy errors on the agent

* Minor cleanup and comments

* Upgrade dependancies for colors, uuid

* Upgraded deps

* HTTP should return a JSON error message when a device does not exist.

* Cleanup domain in test to ensure it does not catch unrelated errors.

* Add --exit to mocha to be compatible with old style in 1.x

Our tests are scheduling async operations and calling done. We
need to clean this up but with the old version of mocha it was
exiting the test suite after all tests were done. This enforces
the old behavior.

* Handle spdy Connection errors in the PeerSocket

* Drop travis support for 0.12

* Remove use of ES6 default parameter.

* Drop support for node v4 in .travis.

* Add osx to travis

* Peer z2z transport always uses spdy/3.1 to support old zetta hubs.

This ensures zetta hubs running pre spdy-upgrade will be compatible
at a network level. To support `h2` we need to negotiate transport
and ALPN cannot help us.

* Upgrade zetta-cluster to fix security vulnerabilities.

* Upgrade spdy to 4.x

* Publish: 1.6.0

---------

Co-authored-by: Adam Magaluk <AdamMagaluk@google.com>